Changed the `aatnode` web application to use FIDIA to get and display the result set.
Instead of passing the query request directly to Spark, it is instead passed to FIDIA, which returns a new FIDIA sample object. This object is then asked for the HTML or CSV version of the tabular data.

(Under the hood, the sample object is taking the Spark DataFrame returned from the query, and converting it to a Pandas DataFrame, which is what FIDIA will use internally, at least to start with).