An attempt at programatically generating SQL using SQLAlchemy.
Unfortunately, to do the SQL using a package that provides some security against SQL injection attacks is proving difficult. The package really needs to be able to identify tables entered in the select part of the statement as the same as those entered in e.g. the join or filter part. This will take some time and effort, and is better addressed under #ASVO-280.

I'm checking this in for safe keeping, but will instead generate the query by hand (and assume that the inputs are sane, i.e. opening the code up to SQL injection attacks).