Query generation from form data largely working.
The query is now generated by hand, with no sanity checking. This will need to change in future for security reasons. However, this seems to work well.

With this commit, the UI can now pass both #ASVO-216 and #ASVO-217. However, it doesn not pass #ASVO-222 (which the UI doesn't support).

A few important notes:
- This code is necessarily tied to the names of the form POST variables, and will have to be updated if those are changed.
- BETWEEN is currently interpreted by splitting on a comma between the two values in the value box. This is fragile: multiple commas will make a mess of the SQL.
- This code should generally be considered fragile. It has not been extensively tested. If data is posted to the form not via the UI, then there is huge potential for unintended side effects.
- No protection from SQL injection
- Results currently rendered as HTML only, no option to return CSV or other options.