Fix for tables with multiple columns of the same name in FIDIA.
The draft of FIDIA assumes that the first column is the index, however, if there are multiple columns with the same name, this makes a mess of the index. I've changed the behaviour to just use an arbitrary index. Side effect is that a new column appears in results named index. A more careful fix is needed.