Trait loading now works as intended.
This was a bit of a challenge to get working!

The values of a trait are marked with the decorator `@trait_property`. When a trait is instantated, the `__init__` function from the higher class calls `self._realise`, which churns through the trait object finding all of the `trait_properties`. It then runs the code at each, and adds the resultant value to the instance's `__dict__`. Python name resolution then ensures that when the value is asked for again, this cached value is returned instead of the original code being called.

`@trait_property` is a descriptor, very similar to the builtin `@property` (actually currently identical, except that the setter and deletter is not implemented). This allows the properties to be cached to be found, and potentially, down the road, would allow writing to traits to run the necessary code to write the data out to the archive.

The caching at the archive level probably needs to be revisited, and possibly redone.