Archives can now provide their own schema.
The schema is returned as a heirarchical dictionary. The top level keys are the "trait types" supported by the archive. Each points to a dictionary which is the schema for the corresponding traits. The dictionaries for the traits have the property name as their key, and the data type (as a string) as their value.

See the test at `test_sami_archive_has_correct_schema` for an example.

The system combines all trait classes that respond to a given trait_type. So if the archive writter decided to write two different trait classes to respond to a trait like "image" (which individually answer to different trait_names, for example), and those trait classes provided different trait_properties (unlikely, but certainly possible), the schema returned would include all of the trait_properties of all of the classes. However, if a particular trait_property is defined with two different types, then there will be an (undefined) error.