Both SAMI traits now have a `known_keys` class method which provides a list of all valid TraitKeys.
This is only half implemented in that I want to be able to ask the archive for this information, but currently you have to ask the individual Trait classes. For SAMI, it is very slow because it basically has to go through all the data to work out what the keys are. Not ideal—we will have to push for the SAMI Team to provide this information in future.