Bug fix for retrieving 0.5" cubes without a version identifier.
The test `test_retrieve_data_from_sample_version_not_required` only tested a single example. It turns out that there was a problem with the way the request for red.05 and blue.05 cubes were retrieved which meant that the code would throw a multiple versions error (with the second version being the red.10). This fixes that by more carefully specifying the wildcard on the plate ID. If the plate ID changes much, particularly if the TILE number ever has more/less than three digits, then this will start failing again.