Django web app still not running. More missing items filled in. (NON-FUNCTIONING)
The `asvo_spark.py` module wasn't linked into the `web_applications/` directory, so that is now done. Also, within astro_spark, the path to import asvo_spark was wrong (pointing inside of FIDIA), so that has been changed as well.

I have temporarily fixed the import error present in the previous commit by disabling a check inside of `slogging`. I think this check is failing because Django also registers handlers, so the assumption that there will be no handlers for a logger may not be valid.

However, this still doesn't work. Spark is throwing an "invalid syntax" error at a non-existant line of the code.

I'm checking this in for safe keeping, but this has to go to the back burner for a while.