SAMI Traits now support serialization via the `.as_bytes()` method.
This is a bit of a test case. I realised that serializing the individual parts of a trait don't make much sense: either the database should be able to understand all of a trait, or none of it. And if the answer is none, then why break it appart rather than just store it as a single blob?

This is implemented to basically pickle a dictionary containing only the trait_properties for a given Trait. This suggests a lot of refactoring under the hood which could make the whole trait retrieval system much clearer.

Basically, the idea is as follows: a Trait maintains a special dictionary _trait_data, which is basically designed to work like __dict__, but not be under the control of python. the trait_property decorator then by default defines the "loader" for a descriptor object. The descriptor's __get__ when called, first checks the _trait_data dictionary for the value, and if it isn't there, then considers running the loader, along with calling any pre-loader/post-loder steps. This greatly clarifies how the properties are cached, as they don't rely on an understanding of python's attribute resolution order. Furthermore, the dictionary can easily be pickled (or otherwise serialized, if we think of better ways), while clearly seperating out things which are not part of the data. If new properties are added to the Trait, then re-loading from the serialised data won't break: the __get__ will simply have to call the loader for new data (of course if the loader has changed, then the old serialized data will have to be invalidated somehow.

I've written a test in the `test_sami_serialization.py` module, that checks that we can at least recover the data in the current implementation.