Changed TraitProperty caching to use a seperate cache dictionary, rather than caching in the instance's `__dict__`.
This is much more transparent programming, as it removes the trait property caching from the details of the python MRO and instead makes it more explicit in the trait property (the decorator now defaults to saving the user provided function as a loader, while the getter takes care of checking the cache.)

Now traits have a `_trait_dict` which is used to create a cached copy of all of the `trait_properties` when the trait is constructed. TraitProperty descriptors also know how to add a new value to the cached copy if it is somehow missing but the trait knows how to load it (as it should).

JIRA Issues: ASVO-377