Fixed the Trait `schema` function such that it correctly includes the schema from inherited classes.
The primary change here is to iterate over the items of `dir(cls)` instead of over the items of `cls.__dict__`. The former is a Python API call which will generate a list of all valid attributes for the given object, while the latter will only list attributes defined for this class (but does not include any parent classes).

See http://stackoverflow.com/questions/14361256/whats-the-biggest-difference-between-dir-and-dict-in-python

Also, this SO explains how we know we'll get the actual descriptor object, rather than calling it as an attribute
http://stackoverflow.com/questions/17735520/determine-if-given-class-attribute-is-a-property-or-not-python-object
(the short answer is that if it is requested through an instance, then `__get__` is called, but if through the class, then the descriptor object is returned.)

The test `test_trait_inherited_has_correct_schema` now passes.