Traits for LZIFU now more fully implemented and tested. Still a ways to go.
Some changes to better handle the case where a TraitKey for non-existant data is provided. I've added a try/except block basically around the whole thing which will throw a DataNotAvailable error unless the underlying code throws a DataNotAvailable error. THIS CAN AND WILL MASK NECESSARY EXCEPTIONS. There is a TODO to go and fix this in the code.

Also updated the `test_sami_archive` and `test_sami_lzifu` to test more of the LZIFU stuff.

Still to go is more complex things, like WCS for the LZIFU maps, better trait inheritance and structuring, etc, etc.

ASVO-330