Refactored Traits to include pointers to parent_trait as part of heirarchical data structure.
I have changed the signature for `Trait` to include `object_id` and `parent_trait`. These arguments are implemented as kwargs to facilitate some backwards compatibility so that I don't have to do as much refactoring immediately, but the `object_id` is effectively required.

The TraitKey has not been refactored as part of this.

I have also refactored the use of `__init__` in the team archive plugins to `init`, which reduces the plug-in's dependence on underlying changes in FIDIA, and is more robust.

This seems to pass tests, and looks like it is working. At this stage, no traits are actually hierarchical. Furthermore, Traits themselves will now have to implement some kind of `get_trait` method similar to what is currently implemented at the Archive/AstroObject level, though it should not be as complicated.

ASVO-334 ASVO-387 ASVO-388 ASVO-348