First attempt at writting a FITS exporter.
The FITS Exporter is at `Trait.as_fits` (to match `Trait.as_bytes`). I have hashed out all of the code to export all of the kinds of things that might be included in a FITS file, but then commented out bits that are not fully implemented/ready.

There are two other new functions to support this code:
`Trait.sub_traits` and `Trait.trait_property_values`, which each basically return an iterator that can be used to retrieve all of the Traits/TraitProperties of the supplied type (or `None` for any type). For TraitProperties, the value of the t.p. is returned, not the descriptor object.

ASVO-384 #time 1h
ASVO-387 ASVO-391