The schema now supports hierarchical traits.
This was harder than I thought: originally, I had envisiaged (mostly because it was the simplest solution) sub-trait classes being identified and added to the trait on instantiation, and not be part of the class. However, we want to be able to recover the schema "up front", which means we can't wait for traits to be instantiated. Therefore sub-trait classes have to be identified on the class, not on instances of the class.

There were a few place-holder functions on `Trait` which are now implemented (in consideration of the above).

The linking of a sub-trait class to it's parent trait class should be reviewed, as it is currently very fragile and will almost certainly be broken by the first user to try to implement them. (see ASVO-397)

Also implemented are some tests that proves all this works. These are implemented at `test_traits`, and take advantage of some new stuff added to the `ExampleArchive` in `archive/test_archive.py`

ASVO-394 #time 1h 30m #done