Working version of JSON GAMA table descriptions, but slow.
This is a first go at ASVO-417, but querying via Hive seems to be very slow. On my machine, this takes many 10s of seconds. Not sure why it is so slow. Manually running some of the queries against pyspark shows that they run faster there, but the overhead still makes them slow. Could this just be that Hive is using MapReduce, and Spark is smarter?

Best plan is probably to just precreate the file and then serve it statically. So just such a file has been added to the rest_api app.

ASVO-417 #time 2h 30m