First attempt at refactoring traits into fully fledged objects.
So, I realise in working through this that the best way to go about doing this is not nearly as clear as I had originally thought. This approach keeps the descriptors, but has them replace themselves with `BoundTraitProperties`, which basically keep pointers to both the original TraitProperty and the object so that the values can be retrieved. In this approach, the syntax of FIDIA changes slightly in that values of TraitProperties accessed from their Traits via attribute must be "called", i.e. they must have parens after their name now (otherwise, the BoundTraitProperty object is returned).

An alternative approach would be to keep the syntax the same, but modify the `Trait` helper function `_get_trait_properties` to return these `BoundTraitProperties instead of the `TraitProperty` descriptors (so they could be asked for their value directly, rather than by having to pass in the object.

Either approach will work for Django Rest Framework. Incidentally, the framework automatically calls the attribute if required, so potentially no refactoring of Django is required (see `rest_framework/fields.py/get_attibute` (about line 65).

This commit includes refactoring of the tests and SAMI archive, and they seem to be passing. The ExampleArchive and ingestion routines have not been updated. Furthermore, I suspect other bugs will have crept in—the tests are actually rather thin on the ground for some of this stuff.

ASVO-425 #time 3h #in-progress