Refactor: CSS to LESS.
Allows for cleaner structure in the style sheets (nesting, variables, scope etc). As more and more endpoints are being developed, managing the styling rules was getting unruly (which surpasses which, forcing with !important etc)

CSS is now split into a global sheet (main.less) that's linked in each page's header and provides the 'feel' of the site. Individual page styling will be in subdir of less/.
Note - if others wish to edit the .less you'll need to install a less transpiler on your IDE.

ASVO-516 #time 3h #in-progress