First working version of the "slow" cache for FIDIA (ASVO-480).
Cache's are effectively to be implemented as a "stack". A request is passed to the top cache on the stack. If it can't answer the request, that cache passes it to the next cache in the stack. When a cache does finally respond, that response is passed back up the stack so that any interveining cache can store the response for future requests. The top cache then finally returns to the calling TraitProperty.

If no cache in the stack can answer the request, the last cache has a pointer to `None` (rather than another cache) in which case it makes the request against the `uncached_value` property of the TraitProperty, which is guaranteed to provide a response.

In this version, there is no machinery to add or remove caches from the stack. A `MemoryCache` is the only item on the stack when an archive is created. This will need to be sorted out. It probably makes sense to keep the list of caches in the stack on the archive, and have each cache somehow refer back to the archive to get the next cache from the list rather than having the archive update each cache object when the cache list changes. Probably depends on the implementation details of `yield` and some testing of how complex the code looks either way.

Most of this is not yet well documented, so I've opened ASVO-517.

A few things have cropped up:
- BoundTraitProperty should probably be re-integrated into TraitProperty, and have a `bind` command or similar. This will probably simplify the code a bit.
- Testing is probably a bit thin on the ground. In writting this, I could imagine several ways in which the code could fail that the tests won't pick up. Specifically, there was an error in `uncached_value` that meant that traits were not being cleaned up after having been loaded. ASVO-518

Working through this has revealed several things about caching I had not fully appreciated when I started:

- There is actually a clear distinction between the "slow" cache and the "fast" cache. Two things can be cached when dealing with FIDIA Traits and TraitProperties: the data, and the python objects themselves. This commit handles the former: caching the actual data.
- It is difficult to decide how a cache can decide whether it should load only a single TraitProperty, or all of the TraitProperties for a Trait. There are a lot of reasons why it can be important to decide between 'lazy' and 'verylazy', but I'm not sure what all of them are or how to address them yet.
-  There are probably a lot of performance considerations. I don't have a feeling for what will be the bottlenecks in this code, but doubt it will be particularly efficient as is (though there are no particular reasons to think it would be inefficient).

ASVO-480 #time 6h