Refactored TraitKeys to now be trait_type, trait_qualifier, branch, version.
This addresses ASVO-387. The TraitKey is now much more specific: it only contains the information to identify the trait within an archive assuming that the parent trait (if any) and object_id are already known separately. In effect, it is defining this particular sub-part of the data model.

Some things still to look at: the TraitMapping class, and it's wildcard notation. This probably needs to be reviewed and/or updated at some point.

This is passing basic tests, but may still have some bugs.

ASVO-387 #time 1h 30m
ASVO-389 #done