Traits now set up to inherit branch and version information if not explicitly provided.
This addresses ASVO-399. Within FIDIA, it is still possible to have branch and version inforation on sub-Traits, but if none is provided, then it simply inherits from the parent trait.

Therefore, the ASVO can be supported with versions and branches filtering down to sub-Traits where they are used at the Astro-Object level.

ASVO-399 #time 30m #done