Fixing an introduced bug from previous commit on ASVO-387.
In the process of refactoring the `object_id` out of the `TraitKey`, I had neglected to realise that there were cases where I did not explicitly include the `object_id` in calls to make a new `Trait`. Bitten by the fact that information could (previously) be provided multiple ways.

Unfortunately, this wasn't picked up by tests, at least not by the basic tests. (SAMI tests might have failed, but not sure). I need to write some tests that confirm that when a `Trait` is requested, it is actually the one I asked for! This is raised in ASVO-541

ASVO-387 #time 30m