Added two higher levels of debugging: vdebug and vvdebug.
These higher levels of debugging are perhaps appropriate for inner loops, etc, that will produce A LOT of debug messages. Mostly implemented to try to cut down on some of the rather verbose debugging in some modules.

These can be called in the same way as the other debug functions:

	log.vdebug(message)
	log.vvdebug(message)

Also, there are new named levels:

	slogging.VDEBUG
	slogging.VVDEBUG