User can ingest data in batches incrementally.
For this to work, already ingested data needs to be available in an Impala table.

This can be run as a python script and following command line arguments and options are required.

-t --table : impala table of the already ingested data
input_dir : Input data directory
catalogue : Input data catalogue
output_file : An avro file path to write the data into

ASVO-560