First go at getting descriptions working, ASVO-536.
This commit implements the various parts of the FIDIA description specification (Pretty Name, description, documentation) as Python descriptors which are added to the class. The descriptors can be initialised with a value (to explicilty provide the values), otherwise they try to look up the value through the .__doc__ or .doc attributes of the class to which they are attached.

This needs to be made smarter: for instance, how far up should the recovery from .__doc__ be allowed to go? It doesn't make any sense to provide the doc-string from the base Trait class as documentation for a particular trait.

Also, needs to have a "short name" added to the spec to handle cases like FITS header keyword names? And we should consider possibly adding a providence (probably part of a bigger providence discussion).

Furthermore, not clear how this will be added to the database, if at all, and hence how will it be versioned.

ASVO-536 #time 3h