Most of FITS meta-data added to SAMI archive, ASVO-571.
FITS meta-data from the Cube headers has been added to the spectral_cube trait and associated sub-traits. To get this working, I had to implement external Trait pre-loading, which is messy. But I think the Traits and sub-Trait design that will solve many of these problems is becoming more clear.

This code is not at all DRY. I think I have some ideas of how I want to fix this, see ASVO-573 and FIDIA/Traits.md on the wiki. But for now I just want to get it working.

ASVO-571 #time 3.5h