Another attempt to make the descriptive data work the way I want.
In this attempt, I thought I could take advantage of the way python descriptors were implemented to keep the descriptor object, but still set to it at the class level. However `__set__` doesn't provide the functionality I thought it did. This code is basically incomplete (and the corresponding test is failing.)

So, I'm going to go a different route for the moment: simply make the descriptors attributes with no smarts (see next commit.)

ASVO-571 #time 45m