Added convenience function to improve DRYness of code.
I've basically implemented a TraitProperty factory, which builds traits which otherwise would look very similar except for a few things.

This is currently in traits/utilities.py, though it would make sense to think about restructing the TraitProperty code out of utilities.py and into its own module.

The situation could be further improved by allowing TraitProperties be put into a list on the Trait, and then the Trait when it initialises could create attribute references to the Trait Properties, and add any manually defined TraitProperties to the list as well. This is heading in the direction of having Traits behaive more like dictionaries (ASVO-520).

ASVO-571 #time 45m