A modified implementation of descriptions using a mixin ASVO-536
This replaces the previous implementation with a mix-in class. The class stores its data as class attributes of the current class, and therefore handles inheritance of the documentation automatically. This might need to be slightly smarter, but certainly should work for the moment.

The mechanism for returning the pretty name will probably need work: how should it decide what to return based on e.g. which qualifier has been requested? This should probably be written while simultaneously working out the Django interface, as there may be interesting constraints provided by Django that would be better addressed in design than retrospectively.

ASVO-536  #time 2h