Modification of the Django serializers to include the descriptions data.
The actual code to do this is commented out!!! (though the supporting code is not).

This has a few issues: it adds fields which FIDIA effectively doesn't know about (at least not in the way that Django expects it to), so for instance one can click on the description field in the list of data, but the url breaks because Django asks FIDIA about something it doesn't understand ("html_documentation" is not a TraitProperty!).

I think the short answer is that we don't want to do this, hence why I've included it commented out. We can discuss.

ASVO-588