Fixing up various parts of the descriptions for SAMI. Several small bug fixes.
Basically looking at how the SAMI descriptions appear in the front end. I've discovered a few bugs along the way:

(1) The biggest was that the trait_qualifier was not being handled correctly, and more importantly requesting a `pretty_name` for a Trait that didn't otherwise have it would fail because it was requesting the trait_name instead of the trait_type. I've extended DescriptionsMixin for Traits. This enables pretty names for particular qualifiers to be defined.
(2) Currently, different documentation for different qualifiers handled by the same Trait class is not possible. For the moment, this is probably best achieved by using different classes where the documentation must differ (if the documentation is different, then chances are the Trait class will need to be different anyway.)
(3) Third bug is that if there was no format provided anywhere for the documentation, then it would fail conversion. So I've now set the default to be assumed to be `markdown`
(4) In python 2, we'll have to be careful about unicode and file encoding.

ASVO-588, ASVO-474