Change the way descriptions are implemented slightly.
I realised there was a snag in the way I had originally written the `DescriptionsMixin`. Descriptions which are to be attached to the class only were working fine (and this is what one wants for Traits). However, they did not work correctly for situations where the description data should actually be attached to an instance, as is the case for TraitProperties.

This is slightly confusing because of the way TraitProperties work within Python/FIDIA. Because they are implemented as descriptors, each TraitProperty is an instance, but that single instance is attached to the class of the Trait, and when an instance of the Trait is created, the TraitProperty instance (descriptor) attached to the corresponding Trait class is used to retrieve data, etc, and in the current implementation, a BoundTraitProperty is created which is particular to this individual Trait on this individual Astronomical Object. So, for the purposes of descriptions, Trait instances are to Trait classes as BoundTraitProperty instances are to TraitProperty instances.

This makes descriptors confusing, because they need to change their behaviour for different kinds of objects they might be attached to. As I see it, there are basically two options:

- Create two DescriptorMixin classes, one for cases where descriptions should be attached to instances and one where they should be attached to classes
- Adapt the same code to work in both cases, but respect a setting on the class mixed into that determines what is allowed.

I've opted for the latter option. This makes the implementation of the DescriptorMixin a bit messy, as it must include the code to check what is allowed, and also because it makes most of the methods both class and instance methods.

I'm not sure I'm happy with this at all, but it does seem to work currently. Making this clean would probably require rethinking both the implementation of FIDIA descriptions and of Traits/TraitProperties in general, which is not worth considering now.

ASVO-474