FIDIA Schema can now return catalog or non-catalog data only as well as the whole schema.
I've implemented a new keyword on the schema functions, `data_class`, which can be one of 'all', 'non-catalog' or 'catalog'. Setting it to one of the later options will restrict the returned schema to only include that kind of data. All other options behaive as before.

Tests in place, and passing.

Next, we'll have to decide how to feed this into the query builder.

ASVO-582 #in-progress #time 2h