Improved logic to determine when Trait is a 2D Map and should be rendered on the web front end with the little viewer.
Basically, I changed the logic from having a hard-coded list of trait_types to instead check if the `Trait` is a subclass of `Map2D`. Not sure this is fully implemented.

Also note that as we get more of these, the template and renderer will probably need to be broken up, but that will require changing where in the code is the logic of determining the which Django View to use.

ASVO-579