Working version of the new schema function in place for Liz.
I've deleted the extra code in serializers.py and views.py. As a result, none of the schema browser chrome appears in this version.

Also changed the default option for combine_levels to be none, rather than an empty tuple, which was confusing and hard to type. Probably worth changing the logic to work with sets or lists instead of tuples anway...