Bug fix in pretty_name logic for ASVO-578.
There was a problem in that the qualifier level was setting a pretty name on the schema, which the Trait.schema would try to reset (to something different) resulting in an error. In this version, the qualifier level doesn't set if there is not going to be a branch and version level below it.

This is messy. Not sure that having all these generic 'pretty_name' items floating around is a good idea.

ASVO-578 #time 45m