Changed descriptions.py such that it won't duplicate a description at the top of a documentation string when pulling the documentation string from the class doc-string.
Basically now checks if the documentation pulled from the docstring starts with the description, and if so, trims it off.

Also added some more tests to descriptions (one of which, that for a description that has been wrapped across more than one line) fails. The description parser needs to have the deindent function added.

ASVO-589 #time 30m