SAMI Public archive now has some featured properties.
The change to the SAMI archive is pretty minimal: only a new list of propeties to feature.

However, I also realised that TraitPaths need to be able to update keys with defaults when asked to retrieve the Trait/Property corresponding to the path.

TraitPath requests for a TraitProperty assume that the default 'value' is meant if no TraitProperty name has been given.

ASVO-692 #time 30m #in-testing