Attempt to fix the bug on filtering catalog/non-catalog data.
I've basically taken advantage of the fact that the 'data_only' version of the schema seems to be correct. When another version is requested, the function also creates the data_only version, and then deletes any keys not found in the 'data_only' version. This is done in the `TraitRegistry` part of the schema generation.

NOTE: this probably doesn't work if there are some levels combined. More testing needed.

I've written a test of the first (in order to develop the solution) but it is against SAMI since I knew how to do that. It should be moved to a generic test on ExampleArchive. ASVO-733

ASVO-735 #ready-for-testing #time 20m