Fix to a strange problem with CatalogCoordinate.
It seems that the way catalog_trait overrides properties is insufficient for ABCMeta to recognise, so it is then impossible to instantate a SkyCoordinate created with catalog_trait. For the moment, I've simply eliminated the abstractproprties to make this work.

Anyways, now passing SAMI tests at least, so hopefully good to go.