Changed LZIFU file finding logic to handle cases with repeated observations.
Before, for the handful of galaxies where there was more than one observation, the LZIFU data would not be found becuase the filenames for these files is different. Now, if the code doesn't fine the file using the standard logic, it tries again using the extended logic for these special filenames.

Also includes a test.