Fixed an error in the logic for inheriting branch names from parents.
Originally, when I wrote this I imagined a special key "Inherit" which would tell a trait that it should just use the same branch/version information as it's parent. Ultimately, the we have made several changes:
- branches and versions must be defined for top level traits.
- sub-traits may not have explicit branch/version information
Therefore, in some sense, the only thing a sub-trait can do, other than have no branch/version info, is to inherit it from it's parent. Unfortunately, the logic for this inheriting was broken.

It remains useful to have branch/version information trickle down to sub-traits as they may want to use it to help with retrieving data.