Fixing a problem with type casting in the SAMI plugin for the LZIFU Flags.
Turns out that np.ndarray.astype(..., casting='safe',...) doesn't do what I understood it to do. It does no check of the actual data, just tells you if the cast requested is garanteed to be safe for any input or not. (kind of silly, really). So I have done the test for the safe cast myself.