Rewrote the SFRMap class in the SAMI plugin to return maps that really are only one component at a time.
I had not previously realised that SFR is computed per component, as well as a total. Therefore, these maps had higher dimensions than was expected. This is now corrected. (However, there is still no test that the dimesionality of a TraitProperty matches that specified in it's type.

ASVO-810 #time 1h

branch: ASVO-810-A