SpectralMap no longer requires variance to be defined.
Requiring variance doesn't make sense in some cases, such as when presenting a fit instead of observed data.

This change also prevents an error when variance isn't defined by a subclass, as was the case in LZIFUCombinedFit. I've now removed the "error" property from that class as well.

ASVO-760 #time 10m

branch: sami-fixups