OII 3726 and 3729 now presented as a single OII3727 doublet.
I have removed the two lines from the single component class, and created a new class LZIFUOneComponent3727 to provide the necessary logic.

This new class overrides the value and sigma properties, returning the sum and sum in quadrature, respectivley, of the two lines. I also had to override _wcs_string since it uses the qualifer to find the necessary extension.

Not yet complete—the same must be implemented for the multi-component fits.

ASVO-763 #time 20m

branch: ASVO-763