Changed SchemaDictionary to be based on a "SortedDictionary" so that some kind of ordering is preserved. NON-FUNCTIONAL
This partially adresses various issues with the ordering of the data in the UI. The idea here is to enforce some kind of order which is at least persistant/consistent, rather than to fully solve the problem. This implementaiton simply makes the SchemaDictionary into a version which is always sorted by it's keys. The front end won't quite appear sorted as the schema is keyed by trait_type, not by the pretty printed version of the trait type. So, for example, "Galactic extinction" appears at the top because it's key is "A_g".

Basically, I've changed SchemaDictionary to subclass SortedDictionary instead of just a plain dict. There are two additional changes required to make this work:

1) Normally, SortedDictionary sorts based on the key. SchemaDictionarys have ints, strings, and None as keys, which are not intersortable. So I use `str` as the sorting key function to SortedDictionary

2) This version won't pass the tests. This seems to have to do with "deepcopy"ing SortedDictionaries. See the next commit for a fix.

ASVO-767 #time 30m

branch: ASVO-767-A