SortedDictionary ordering of SchemaDictionaries now passes tests (and works!).
There seems to be an issue with "deepcopy"ing SortedDictionaries, at least in the subclassed SchemaDictionary.

This commit is effectively a work around that removes the relyance on `deepcopy`. I've made it a separate commit, as it ultimately would make more sense to solve/address the underlying problem, but not now.

ASVO-767 #time 45m

branch: ASVO-767-A