Update of the Multi-component velocity maps to now include all components.
I've done this in a strange way: The value of the trait now points to the 3D stack of velocity maps by component. This is visualised with Liz's data source selector. The individual component maps are then added as separate TraitProperties.

This is not ideal. Individual maps should probably be Traits, but anyway. If there was ever a case for a `ListTrait` then I think this is one of them.

ASVO-775 #time 30m

branch: ASVO-775-A