Improvements to unit handling so that they can be bubbled up to the Web frontend.
These changes are largely to address ASVO-757.

There is a new function on a Trait class, `get_formatted_units`, which attempts to format the `astropy.units.Unit` or `astropy.units.Quantity` stored at `Trait.unit`. I suspect this is going to need improving based on feedback from Liz.

To support this new function, units must always be stored as class attributes, not properties/descriptors. So I have changed the way `catalog_trait` sets up the unit attribute.

`Trait.full_schema` then uses the new `get_formatted_units` to add a unit item to the schema for a trait for the levels of 'metadata' and 'descriptions'.

Also, `get_feature_catalog_data` on `Sample` now also uses this same function to populate an additional item returned so that Liz can use it to format the content of her tables.

ASVO-757 #time 1h

branch: ASVO-757-A