Dimensionality specified in TraitProperty type did not match data for some data.
velocity_map-ionized_gas:recom_comp and velocity_dispersion_map-ionized_gas:recom_comp both had type "float.array.2", but are actually three dimensional. These really shouldn't be three dimensional, but I have left them in with higher dimensions so that those traits can still have a sensible "value" TraitProperty. This may not be the best plan, as it leads to duplication of the data in the system, but it is the same data.

branch: master