Finally adding the hyphenation only version of the string trait key.
Liz and I discussed this at length to handle issues with the front end and anchor names. We came up with an alternate notation that avoided some reserved characters to ease handling on the front end, but then it turned out not to be needed. I'm implementing it now as I need it to handle file name construction for ASVO-870 (as both paren and colon are not pretty in filenames).

ASVO-870 #time 30m

branch: ASVO-870-A