First go at having FIDIA produce a latex report of an archive's contents.
I've create a new module, reports, which contains the necessary code. Basically the code works over an archive's TraitRegistry, producing a LaTeX report on the contents of it.
This does not yet correcly handle sub-traits (better recursion is necessary). It does, however, include the source-code of the relevant parts of the team plugin. LaTeX formating could be dramatically improved: this is quite verbose and isn't very compact on the page.

Better integration with pylatex would also make the code easier to maintain and extend.

branch: fidia-content-report