Fixed an issue in BoundTraitProperty attribute lookup that was preventing descriptions from working correctly.
This is slightly messy. We need to be able to call `get_description`, etc, on BoundTraitProperty and have it work as expected. `get_description` and its relatives store information on the instance (in this case) of the TraitProperty. Therefore when called on the BoundTraitProperty, the requests for the `_short_description` attribute, etc, must be forwarded to the corresponding TraitProperty. This wasn't happening because the getattr forwarding function was not forwarding requests for private variables.

I've changed this now so it only won't forward requests for system attributes (i.e. those starting with a double underscore).

Tests updated to check for description results explicitly.

branch: ASVO-934-a